Drop depth from params, #define representation
Instead of specifying the display depth of characters via a constructor
parameter, #define constants and set them in the constructor explicitly.
The depth can be overridden via a setter.

Also move the representation of different entities to a single header file
and move the representation constructor param to last position.

Drop hostility flag, as that will be set via bit masks later.