Add a simple message queue
Entities / the game will be able to push messages to the queue. These
messages will be then displayed successively on the screen.

This is still missing splitting the message on spaces, rather then by
fixed chunks. Also, it might be a good idea to store just pointers,
instead of whole strings, to save memory.